Pas grand chose mais on est la

+++ src/info_zone/info_zone.cpp
@@ -21,7 +21,7 @@ InfoZone::~InfoZone(){
     delete bg;
 }
 
-int InfoZone::get_value(){
+float InfoZone::get_value(){
     return value;
 }
